Use Arrays rather then ByteBuffers for JNI calls
ByteBuffers are very slow in Java. For the readRegister and writeRegister, the byte[] calls are faster then the ByteBuffer calls at length 2.

For the auto SPI call, at your length the int[] call is slower, but you are accessing lots of bytes in the Java side.
The Java side array read is exponentially faster, and is worth it over the bytebuffer call.